Tool 1 en 2 compleet, Tool 3 versie 2 aangepast

+++ Straatmodel/SchrijfBestand.cs
@@ -8,34 +8,34 @@ namespace Tool1_BestandSchrijven
 {
     class SchrijfBestand
     {
+        /*We overlopen per straat de graaf, en voegen gaandeweg de knopen en segmenten die we tegenkomen op. 
+            * MAAR:
+            * - We zullen segmenten verschillende keren tegenkomen in 1 straat, aangezien ze in de graaf aan meerdere knopen gekoppeld zijn
+            * - We zullen over de straten heen verschillende keren dezelfde knopen tegenkomen omdat deze knopen de verschillende straten verbinden
+            * 
+            * Dus ik hou van beide bij welke er al uitgeprint zijn! 
+            * */
         public static void PrintDocumenten(Dictionary<string, Dictionary<string, List<Straat>>> provincies)
         {
-            if (File.Exists(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Straten.txt") && File.Exists(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Segmenten.txt") && File.Exists(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Knopen.txt"))
+            if (File.Exists(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Straten.txt") && File.Exists(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Segmenten.txt") && File.Exists(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Knopen.txt") && File.Exists(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Punten.txt"))
             {
                 File.Delete(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Straten.txt");
                 File.Delete(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Segmenten.txt");
                 File.Delete(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Knopen.txt");
+                File.Delete(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Punten.txt");
             }
-
-
-            /*We overlopen per straat de graaf, en voegen gaandeweg de knopen en segmenten die we tegenkomen op. 
-             * MAAR:
-             * - We zullen segmenten verschillende keren tegenkomen in 1 straat, aangezien ze in de graaf aan meerdere knopen gekoppeld zijn
-             * - We zullen over de straten heen verschillende keren dezelfde knopen tegenkomen omdat deze knopen de verschillende straten verbinden
-             * 
-             * Dus ik hou van beide bij welke er al uitgeprint zijn! 
-             * */
+           
             List<int> uitgeprinteSegmentenID = new List<int>();
             List<int> uitgeprinteKnopenID = new List<int>();
 
-
-            //Beter apart of samen? 
             using StreamWriter writer = File.CreateText(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Straten.txt");
             using StreamWriter writer2 = File.CreateText(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Segmenten.txt");
             using StreamWriter writer3 = File.CreateText(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Knopen.txt");
+            using StreamWriter writer4 = File.CreateText(@"C:\Users\Sieglinde\Documents\Programmeren\Labo_Straatmodel\Punten.txt");
             writer.WriteLine("StraatID;Straatnaam;Gemeente;Provincie;GraafID;Lengte");
-            writer2.WriteLine("SegmentID;BeginknoopID;EindknoopID;straatID;puntenlijst");
+            writer2.WriteLine("SegmentID;BeginknoopID;EindknoopID;straatID");
             writer3.WriteLine("KnoopID;X;Y");
+            writer4.WriteLine("X,Y,SegmentID, positie");
             foreach (KeyValuePair<string, Dictionary<string, List<Straat>>> provincie in provincies)
             {
                 foreach (KeyValuePair<string, List<Straat>> gemeente in provincie.Value)
@@ -59,20 +59,14 @@ namespace Tool1_BestandSchrijven
                             {
                                 if (!uitgeprinteSegmentenID.Contains(segment.SegmentID))
                                 {
-                                    string punten = "";
                                     List<Punt> puntenlijst = segment.Vertices;
 
                                     //De eerste en laatste waarde van de vertices knip ik er hier af, die zitten al in mijn tekstbestand van begin en eindknoop!!
                                     for (int i = 1; i < puntenlijst.Count-1; i++)
                                     {
-                                        //Deze if zodat er enkel tussen de 
-                                        if (!(i == segment.Vertices.Count -2))
-                                            punten += $"{puntenlijst[i].X} {puntenlijst[i].Y},";
-                                        else
-                                            punten += $"{puntenlijst[i].X} {puntenlijst[i].Y}";
-
+                                        writer4.WriteLine($"{puntenlijst[i].X};{puntenlijst[i].Y};{segment.SegmentID};{i}");
                                     }
-                                    writer2.WriteLine($"{segment.SegmentID};{segment.BeginKnoop.KnoopID};{segment.EindKnoop.KnoopID};{straat.StraatID};{punten}");
+                                    writer2.WriteLine($"{segment.SegmentID};{segment.BeginKnoop.KnoopID};{segment.EindKnoop.KnoopID};{straat.StraatID}");
                                     uitgeprinteSegmentenID.Add(segment.SegmentID);
                                 }
